third_party/blink/web_tests/fast/forms/text/pattern-expected.txt

CONSOLE ERROR: Pattern attribute value [(] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[(]/v: Invalid character in character class
CONSOLE ERROR: Pattern attribute value [)] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[)]/v: Invalid character in character class
CONSOLE ERROR: Pattern attribute value [[] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[[]/v: Unterminated character class
CONSOLE ERROR: Pattern attribute value [{] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[{]/v: Invalid character in character class
CONSOLE ERROR: Pattern attribute value [}] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[}]/v: Invalid character in character class
CONSOLE ERROR: Pattern attribute value [/] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[/]/v: Invalid character in character class
CONSOLE ERROR: Pattern attribute value [-] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[-]/v: Invalid character in character class
CONSOLE ERROR: Pattern attribute value [|] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[|]/v: Invalid character in character class
CONSOLE ERROR: Pattern attribute value [&&] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[&&]/v: Invalid set operation in character class
CONSOLE ERROR: Pattern attribute value [!!] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[!!]/v: Invalid set operation in character class
CONSOLE ERROR: Pattern attribute value [##] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[##]/v: Invalid set operation in character class
CONSOLE ERROR: Pattern attribute value [$$] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[$$]/v: Invalid set operation in character class
CONSOLE ERROR: Pattern attribute value [%%] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[%%]/v: Invalid set operation in character class
CONSOLE ERROR: Pattern attribute value [**] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[**]/v: Invalid set operation in character class
CONSOLE ERROR: Pattern attribute value [++] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[++]/v: Invalid set operation in character class
CONSOLE ERROR: Pattern attribute value [,,] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[,,]/v: Invalid set operation in character class
CONSOLE ERROR: Pattern attribute value [..] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[..]/v: Invalid set operation in character class
CONSOLE ERROR: Pattern attribute value [::] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[::]/v: Invalid set operation in character class
CONSOLE ERROR: Pattern attribute value [;;] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[;;]/v: Invalid set operation in character class
CONSOLE ERROR: Pattern attribute value [<<] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[<<]/v: Invalid set operation in character class
CONSOLE ERROR: Pattern attribute value [==] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[==]/v: Invalid set operation in character class
CONSOLE ERROR: Pattern attribute value [>>] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[>>]/v: Invalid set operation in character class
CONSOLE ERROR: Pattern attribute value [??] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[??]/v: Invalid set operation in character class
CONSOLE ERROR: Pattern attribute value [@@] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[@@]/v: Invalid set operation in character class
CONSOLE ERROR: Pattern attribute value [``] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[``]/v: Invalid set operation in character class
CONSOLE ERROR: Pattern attribute value [~~] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[~~]/v: Invalid set operation in character class
CONSOLE ERROR: Pattern attribute value [_^^] is not a valid regular expression: Uncaught SyntaxError: Failed to read the 'valid' property from 'ValidityState': Invalid regular expression: /[_^^]/v: Invalid set operation in character class
This is a testharness.js-based test.
Harness: the test ran to completion.
